Make udpdateRecipesFromModal dynamic
